layout is beginning to take form
