+++ angular-app/src/app/atmos/apps-container/apps-container.component.css
@@ -1,0 +1,7 @@
+.upload {
+  /* width: 35%; */
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  cursor: pointer;
+}
