+++ angular-app/src/app/atmos/apps-container/apps-container.component.html
@@ -1,1 +1,14 @@
-<router-outlet></router-outlet>
+<p>Apps Container Works</p>
+
+<button mat-raised-button (click)="detectApp()">
+  Curriculum App
+</button>
+
+<!-- <div class="upload">
+  <label id="fileLabel" class="file-btn green circular-btn">
+    <img src="../../../assets/assignment-white-48dp.svg">
+<hr>
+Curriculum
+</label>
+<input style="display: none" id="curriculum" (click)="detectApp()">
+</div> -->
