+++ angular-app/src/app/atmos/atmos.module.ts
@@ -9,7 +9,7 @@ import { MatInputModule } from "@angular/material/input";
 import { MatOptionModule } from "@angular/material/core";
 import { MatSelectModule } from "@angular/material/select";
 import { SharedModule } from '../shared/shared.module';
-import { AppsContainerComponent } from './apps-container/apps-container.component';
+import { AppsContainerComponent } from '../atmos-apps/apps-container/apps-container.component';
 
 // const routes: Routes = [
 //   { path: 'apps', component: AppsContainerComponent }
