+++ angular-app/src/app/layouts/home-layout/home-layout.component.css
@@ -1,0 +1,9 @@
+:host {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+}
+
+.container {
+  padding: 20px
+}
