+++ angular-app/src/app/layouts/home-layout/home-layout.component.html
@@ -1,10 +1,12 @@
-<app-header></app-header>
+<app-header flexLayout="column"></app-header>
 
-<div>
-  <div>Sidebar Works!</div>
-  <div>
+<div flexLayout="row" fxFlex>
+  <div fxFlex="20">
+    <app-sidebar fxFlex></app-sidebar>
+  </div>
+  <div fxFlex="80" class="container">
     <router-outlet></router-outlet>
   </div>
 </div>
 
-<app-footer></app-footer>
+<app-footer flexLayout="column"></app-footer>
