+++ angular-app/src/app/shared/comnponents/footer/footer.component.css
@@ -1,0 +1,4 @@
+footer {
+  padding: 20px;
+  background-color: green
+}
