+++ angular-app/src/app/shared/comnponents/header/header.component.css
@@ -1,0 +1,4 @@
+header {
+  padding: 20px;
+  background-color: pink
+}
