+++ angular-app/src/app/shared/comnponents/header/header.component.html
@@ -1,8 +1,6 @@
-<header>
+<header fxLayoutAlign="space-between center">
   <div>LOGO</div>
-  <div>
-    <a routerLink="/">HOME</a>
-    <a routerLink="/courses">Courses</a>
-    <!-- <a routerLink="/">HOME</a> -->
+  <div fxLayout="row" fxLayoutGap="20">
+    <a routerLink="/home">HOME</a>
   </div>
 </header>
