+++ angular-app/src/app/shared/comnponents/sidebar/sidebar.component.css
@@ -1,0 +1,4 @@
+* {
+  padding: 20px;
+  background-color: purple;
+}
